requirements_no_git.txt
absl-py==2.1.0
accelerate==0.28.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
bert-score==0.3.13
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
cffi==1.16.0
chumpy==0.70
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
datasets==2.18.0
decorator==4.4.2
deepspeed==0.14.4
diffusers==0.26.3
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
fairscale==0.4.13
fastapi==0.110.0
ffmpy==0.3.2
fonttools==4.49.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
gradio==4.19.2
gradio_client==0.10.1
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.24.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
immutabledict==4.1.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
jiter==0.5.0
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.2
lightning-utilities==0.10.1
llvmlite==0.42.0
Markdown==3.5.2
markdown-it-py==3.0.0
matplotlib==3.8.3
mdurl==0.1.2
mkl-service==2.4.0
more-itertools==10.2.0
moviepy==1.0.3
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
natsort==8.4.0
ninja==1.11.1.1
numba==0.59.0
numpy==1.24.4
nvidia-ml-py==12.535.161
nvitop==1.3.2
omegaconf==2.3.0
openai==1.42.0
opencv-python==4.9.0.80
orjson==3.9.15
pandas==2.2.1
peft==0.10.0
pooch==1.8.1
preshed==3.0.9
proglog==0.1.10
protobuf==4.25.3
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
pyglet==2.0.10
pyparsing==3.1.1
pyrender==0.1.45
python-dateutil==2.8.2
python-multipart==0.0.9
pytorch-lightning==2.2.0.post0
pytz==2024.1
referencing==0.33.0
regex==2023.12.25
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.2
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.40.6
setproctitle==1.3.3
shapely==2.0.3
shellingham==1.5.4
shortuuid==1.0.11
shtab==1.7.1
smart-open==6.4.0
smmap==5.0.1
smplx==0.1.28
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
transformers==4.44.1
triangle==20220202
trimesh==3.9.24
trl==0.8.1
typer==0.9.0
typing_extensions==4.12.2
tyro==0.7.3
tzdata==2024.1
uvicorn==0.27.1
wandb==0.16.3
wasabi==1.1.2
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.1
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0
